bakall/backend26.1.8/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
requests
openpyxl
apscheduler
